Moving discovery.FindAllTargets to srvtopo.
This belongs to srvtopo, and removes the dependency between the two
packages. I want to eventually have the dependency the other way around.
